+++ scanpipe/pipes/scancode.py
@@ -312,9 +312,7 @@ def get_virtual_codebase(project, input_location):
     temp_path = project.tmp_path / "scancode-temp-resource-cache"
     temp_path.mkdir(parents=True, exist_ok=True)
 
-    return VirtualCodebase(
-        location=input_location, temp_dir=str(temp_path), max_in_memory=0
-    )
+    return VirtualCodebase(input_location, temp_dir=str(temp_path), max_in_memory=0)
 
 
 def create_codebase_resources(project, scanned_codebase):
@@ -325,21 +323,25 @@ def create_codebase_resources(project, scanned_codebase):
     CodebaseResource objects as the existing objects (based on the `path`) will be
     skipped.
     """
-    for scanned_resource in scanned_codebase.walk():
+    for scanned_resource in scanned_codebase.walk(skip_root=True):
         resource_data = {}
 
         for field in CodebaseResource._meta.fields:
+            # Do not include the path as provided by the scanned_resource since it
+            # includes the "root". The `get_path` method is used instead.
+            if field.name == "path":
+                continue
             value = getattr(scanned_resource, field.name, None)
             if value is not None:
                 resource_data[field.name] = value
 
         resource_type = "FILE" if scanned_resource.is_file else "DIRECTORY"
         resource_data["type"] = CodebaseResource.Type[resource_type]
+        resource_path = scanned_resource.get_path(strip_root=True)
 
-        path = resource_data.pop("path")
         CodebaseResource.objects.get_or_create(
             project=project,
-            path=path,
+            path=resource_path,
             defaults=resource_data,
         )
 
@@ -350,12 +352,13 @@ def create_discovered_packages(project, scanned_codebase):
     object to the DB as DiscoveredPackage of `project`.
     Relate package resources to CodebaseResource.
     """
-    for scanned_resource in scanned_codebase.walk():
+    for scanned_resource in scanned_codebase.walk(skip_root=True):
         scanned_packages = getattr(scanned_resource, "packages", [])
         if not scanned_packages:
             continue
 
-        cbr = CodebaseResource.objects.get(project=project, path=scanned_resource.path)
+        scanned_resource_path = scanned_resource.get_path(strip_root=True)
+        cbr = CodebaseResource.objects.get(project=project, path=scanned_resource_path)
 
         for scan_data in scanned_packages:
             discovered_package = pipes.update_or_create_package(project, scan_data)
@@ -370,7 +373,7 @@ def create_discovered_packages(project, scanned_codebase):
             )
             for scanned_package_res in scanned_package_resources:
                 package_cbr = CodebaseResource.objects.get(
-                    project=project, path=scanned_package_res.path
+                    project=project, path=scanned_package_res.get_path(strip_root=True)
                 )
                 set_codebase_resource_for_package(
                     codebase_resource=package_cbr, discovered_package=discovered_package